Crate: extract backend-agnostic paint scene into takumi-core
Move the stacking-context build — the walk that decides paint order, z-order
buckets, grouping, and bounds — out of takumi-raster into a new takumi-core
`scene` module. The raster backend now consumes `StackingContextNode` from core
instead of owning the build; the paint half (canvas, subcanvas, draw_*) stays in
raster. Relocate `BlurType` to takumi-css next to `Filter`.

The build half was already backend-neutral (NodePaint carries path/transform/
bounds, not pixels), so this is a pure relocation: golden webps are byte-
identical. It sets up the seam a second backend can consume.
