+++ takumi-css/src/style/properties/filter.rs
@@ -61,6 +61,36 @@ pub enum Filter {
   DropShadow(TextShadow),
 }
 
+/// How a CSS blur radius maps to a Gaussian σ, which sets how far the blur
+/// spreads past its box.
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum BlurType {
+  /// CSS `filter: blur()` — radius equals σ (standard deviation).
+  Filter,
+  /// CSS `box-shadow` / `text-shadow` blur — radius equals 2σ.
+  Shadow,
+}
+
+impl BlurType {
+  /// Maps a CSS blur radius to its Gaussian σ.
+  #[inline]
+  pub fn to_sigma(self, css_radius: f32) -> f32 {
+    match self {
+      BlurType::Filter => css_radius,
+      BlurType::Shadow => css_radius * 0.5,
+    }
+  }
+
+  /// Multiplier from radius to the extent the blur visibly reaches.
+  #[inline]
+  pub fn extent_multiplier(self) -> f32 {
+    match self {
+      BlurType::Filter => 3.0,
+      BlurType::Shadow => 1.5,
+    }
+  }
+}
+
 /// A list of filter operations
 pub type Filters = Vec<Filter>;
