+++ takumi-raster/src/components/blur.rs
@@ -4,32 +4,7 @@ const BLUR_DOWNSAMPLE_TARGET_SIGMA: f32 = 6.0;
 const BLUR_DOWNSAMPLE_MIN_DIMENSION: u32 = 128;
 const BLUR_DOWNSAMPLE_MAX_SCALE: u32 = 8;
 
-/// Specifies the type of blur operation, which affects how the CSS radius is interpreted.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub(crate) enum BlurType {
-  /// CSS `filter: blur()` - radius equals σ (standard deviation).
-  Filter,
-  /// CSS `box-shadow` / `text-shadow` blur - radius equals 2σ.
-  Shadow,
-}
-
-impl BlurType {
-  #[inline]
-  pub(crate) fn to_sigma(self, css_radius: f32) -> f32 {
-    match self {
-      BlurType::Filter => css_radius,
-      BlurType::Shadow => css_radius * 0.5,
-    }
-  }
-
-  #[inline]
-  pub(crate) fn extent_multiplier(self) -> f32 {
-    match self {
-      BlurType::Filter => 3.0,
-      BlurType::Shadow => 1.5,
-    }
-  }
-}
+pub(crate) use takumi_core::layout::style::BlurType;
 
 #[derive(Clone, Copy)]
 struct BlurPassParams {
