+++ takumi-raster/src/path.rs
@@ -1,12 +1,12 @@
 use svgtypes::{SimplePathSegment, SimplifyingPathParser};
-use taffy::{Point, Size};
+use taffy::Size;
 use tiny_skia::{
   FillRule as TinyFillRule, LineCap as TinyLineCap, LineJoin as TinyLineJoin, Path as TinyPath,
   PathBuilder as TinyPathBuilder, PathSegment as TinyPathSegment, Point as TinyPoint,
   Rect as TinyRect, Stroke as TinyStroke, StrokeDash as TinyStrokeDash,
 };
 
-use crate::layout::style::{Affine, LineJoin};
+use crate::layout::style::LineJoin;
 
 #[derive(Debug, Clone, Copy, PartialEq, Default)]
 pub(crate) enum Fill {
@@ -201,45 +201,6 @@ impl Placement {
       self.bottom().clamp(0, size.height as i32),
     )
   }
-}
-
-pub(crate) fn transformed_rect_extents(
-  origin: Point<f32>,
-  size: Size<f32>,
-  transform: Affine,
-) -> Option<(f32, f32, f32, f32)> {
-  let corners = [
-    transform.transform_point(origin),
-    transform.transform_point(Point {
-      x: origin.x + size.width,
-      y: origin.y,
-    }),
-    transform.transform_point(Point {
-      x: origin.x,
-      y: origin.y + size.height,
-    }),
-    transform.transform_point(Point {
-      x: origin.x + size.width,
-      y: origin.y + size.height,
-    }),
-  ];
-
-  let mut min_x = f32::INFINITY;
-  let mut min_y = f32::INFINITY;
-  let mut max_x = f32::NEG_INFINITY;
-  let mut max_y = f32::NEG_INFINITY;
-  for point in corners {
-    min_x = min_x.min(point.x);
-    min_y = min_y.min(point.y);
-    max_x = max_x.max(point.x);
-    max_y = max_y.max(point.y);
-  }
-
-  if !min_x.is_finite() || !min_y.is_finite() || !max_x.is_finite() || !max_y.is_finite() {
-    return None;
-  }
-
-  Some((min_x, min_y, max_x, max_y))
 }
 
 pub(crate) type Command = TinyPathSegment;
